instagram-scraper-unified/requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.0
tenacity==8.2.3

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Logging
colorlog==6.7.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Redis for queue management (optional)
redis==5.0.1

# Scheduling (for local testing)
schedule==1.2.0

# Performance monitoring
psutil==5.9.6

# Type hints
typing-extensions==4.9.0